+++ packages/opentelemetry-resources/src/detectors/EnvDetectorSync.ts
@@ -16,7 +16,7 @@
 
 import { diag } from '@opentelemetry/api';
 import { getEnv } from '@opentelemetry/core';
-import { SemanticResourceAttributes } from '@opentelemetry/semantic-conventions';
+import { SEMRESATTRS_SERVICE_NAME } from '@opentelemetry/semantic-conventions';
 import { Resource } from '../Resource';
 import { DetectorSync, ResourceAttributes } from '../types';
 import { ResourceDetectionConfig } from '../config';
@@ -70,7 +70,7 @@ class EnvDetectorSync implements DetectorSync {
     }
 
     if (serviceName) {
-      attributes[SemanticResourceAttributes.SERVICE_NAME] = serviceName;
+      attributes[SEMRESATTRS_SERVICE_NAME] = serviceName;
     }
 
     return new Resource(attributes);