+++ packages/opentelemetry-resources/src/platform/browser/ServiceInstanceIdDetectorSync.ts
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-export * from './default-service-name';
-export * from './HostDetector';
-export * from './OSDetector';
-export * from './HostDetectorSync';
-export * from './OSDetectorSync';
-export * from './ProcessDetector';
-export * from './ProcessDetectorSync';
-export * from './ServiceInstanceIdDetectorSync';
+import { noopDetectorSync } from '../../detectors/NoopDetectorSync';
+
+/**
+ * @experimental
+ */
+export const serviceInstanceIdDetectorSync = noopDetectorSync;